Add differential testing harness
This runs wgsl shaders and rust shaders and compares the output.
If the output differs, the test fails. Differential testing is better
than snapshot testing or golden file testing as there are no reference
files to get outdated.

Note that we are only using wgpu for now, but I want to also vary
the host-side crate / use `ash`. At least we are using `wgpu`'s vulkan
support on linux, so we are not just testing naga's translation.