Simplify the ash runner.
The ash runner is over-engineered: it supports multiple pipelines, each
one with a separate vertex/fragment shader pair, but it only ever uses
one pipeline. The other runners don't support multiple pipelines, so
this just seems like unnecessary complexity. This is bad because the
examples are currently some of the best rust-gpu documentation there is.

More specifically:
- `rebuild_pipelines` is now just `rebuild_pipeline`, and
  `build_pipelines` is no longer needed.
- `compile_shaders` returns a pair instead of a `Vec`: one element for
  the vertex shader data, and one for the fragment shader data.
- RenderBase no longer needs `Vec` and `HashMap` elements, it can just
  store a single pipeline, a single vertex shader module, and a single
  fragment shader module.
- The following types are no longer needed: `SpvFile`,
  `VertextShaderEntryPoint`, `FragmentShaderEntryPoint`.
- Less iteration and `collect`ing in general.